AI/pyproject.toml
[project]
name = "MCP-MULTI-AGENT"
version = "0.1.0"
description = "MCP활용 프로젝트입니다."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv",
    "langchain",
    "langchain-openai",
    "langgraph",
    "langchain-mcp-adapters",
    "langchain-community",
    "langchain-elasticsearch",
    "mcp[cli]",
    "pandas",
    "numpy",
    "nest_asyncio",
    "fastapi",
    "uvicorn",
    "pyyaml",
    "sqlalchemy",
    "aiomysql",
    "cryptography",
    "pymysql",
    "sqlalchemy"
]

[tool.uv]
index-url = "https://pypi.org/simple"